Hooked up stakign controller Refactored some code
