+++ src/lib/providers/MintProvider.tsx
@@ -3,14 +3,13 @@ import React, { useState, useEffect } from 'react';
 
 //- Hook Imports
 import useNotification from 'lib/hooks/useNotification';
-import { useZnsContracts } from 'lib/contracts';
-import { createDomainMetadata } from 'lib/utils';
 import { NftParams } from 'lib/types';
+import { useBasicController } from '../hooks/useBasicController';
 
 export const MintContext = React.createContext({
 	minting: [{}],
 	minted: [{}],
-	mint: (nft: NftParams) => {},
+	mint: async (nft: NftParams) => {},
 });
 
 type MintProviderType = {
@@ -24,56 +23,24 @@ const MintProvider: React.FC<MintProviderType> = ({ children }) => {
 	const [finishedMinting, setFinishedMinting] = useState<NftParams | null>(
 		null,
 	);
-	const basicController = useZnsContracts()?.basicController;
+	const { registerSubdomain } = useBasicController();
 
-	const mint = (nft: NftParams) => {
-		if (!basicController) {
-			console.error('no controller');
+	const mint = async (nft: NftParams) => {
+		const tx = await registerSubdomain(nft);
+		if (!tx) {
+			addNotification('Encountered an error while attempting to mint.');
 			return;
 		}
 
-		const userHasSubmitted = new Promise<void>((resolve, reject) => {
-			const doMint = async () => {
-				try {
-					// get metadata uri
-					const metadataUri = await createDomainMetadata({
-						image: nft.image,
-						name: nft.name,
-						story: nft.story,
-					});
+		addNotification(`Started minting ${nft.name}`);
+		setMinting([...minting, nft]);
 
-					// register subdomain
-					const tx = await basicController.registerSubdomainExtended(
-						nft.parent,
-						nft.domain,
-						nft.owner,
-						metadataUri,
-						0,
-						nft.locked,
-					);
+		const finishMinting = async () => {
+			await tx.wait();
+			setFinishedMinting(nft);
+		};
 
-					resolve();
-
-					setMinting([...minting, nft]);
-					addNotification(`Started minting ${nft.name}.`);
-
-					await tx.wait();
-					setFinishedMinting(nft);
-				} catch (e) {
-					if (e.message || e.data) {
-						console.error(`failed to mint: ${e.data} : ${e.message}`);
-					}
-					console.error(e);
-
-					reject();
-					addNotification('Encountered an error while attempting to mint.');
-				}
-			};
-
-			doMint();
-		});
-
-		return userHasSubmitted;
+		finishMinting();
 	};
 
 	// TODO: Change this hook to run when minting finishes
@@ -89,7 +56,7 @@ const MintProvider: React.FC<MintProviderType> = ({ children }) => {
 	const contextValue = {
 		minting,
 		minted,
-		mint: (nft: NftParams) => mint(nft),
+		mint,
 	};
 
 	return (
